Made `get_sha_algorithm` return a `Result<MessageDigest, Error>` use `Error::InvalidAlgorithm` in place of panic
Even though you would want to make sure people are using the correct algorithm, panicking while within a `Result<T, Error>` does not make much sense when you could return an error instead than to panic, which may more issues. Having `get_sha_algorithm` return a `Result` is a breaking change if this is ever used either directly or upstream, but it is worth the change than to implement a panic.